agregar2

+++ src/main/java/com/example/demo/Calculadora.java
@@ -2,31 +2,39 @@ package com.example.demo;
 
 public class Calculadora {
 
-	private int num1;
-	private int num2;
+	private float num1;
+	private float num2;
 	private char operador;
 	private float resultado;
-	public int getNum1() {
+	
+	public float getNum1() {
 		return num1;
 	}
-	public void setNum1(int num1) {
+	
+	public void setNum1(float num1) {
 		this.num1 = num1;
 	}
-	public int getNum2() {
+	
+	public float getNum2() {
 		return num2;
 	}
-	public void setNum2(int num2) {
+	
+	public void setNum2(float num2) {
 		this.num2 = num2;
 	}
+	
 	public char getOperador() {
 		return operador;
 	}
+	
 	public void setOperador(char operador) {
 		this.operador = operador;
 	}
+	
 	public float getResultado() {
 		return resultado;
 	}
+	
 	public void setResultado(float resultado) {
 		this.resultado = resultado;
 	}
